ref: migrated relevant files over to dev router, integrated into app.js
