+++ api/app.js
@@ -35,6 +35,7 @@ const gameRouter = require('./game/gameRouter');
 const resetRouter = require('./reset/resetRouter');
 const leadBoard = require('./leaderboard/leadboardRouter');
 const achievements = require('./Achievements/achieveRouter');
+const dev = require('./dev/devModeRouter');
 
 const app = express();
 
@@ -75,6 +76,7 @@ app.use('/game', gameRouter);
 app.use('/reset', resetRouter);
 app.use('/leaderboard', leadBoard);
 app.use('/achievements', achievements);
+app.use('/dev', dev);
 
 // catch 404 and forward to error handler
 app.use(function (req, res, next) {
